# server/app/swagger_spec.json
{
  "swagger": "2.0",
  "info": {
    "title": "Streamix API",
    "description": "API para plataforma de videos tipo YouTube",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Token JWT. Ejemplo: Authorization: Bearer <token>"
    }
  },
  "security": [{"Bearer": []}],
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": ["Autenticacion"],
        "summary": "Registrar nuevo usuario",
        "description": "Crea una nueva cuenta de usuario y retorna token JWT",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [{
          "in": "body",
          "name": "body",
          "required": true,
          "schema": {
            "type": "object",
            "required": ["username", "email", "password"],
            "properties": {
              "username": {"type": "string", "example": "usuario123"},
              "email": {"type": "string", "example": "correo@ejemplo.com"},
              "password": {"type": "string", "example": "password123"},
              "avatar_url": {"type": "string"},
              "bio": {"type": "string"}
            }
          }
        }],
        "responses": {
          "201": {"description": "Usuario creado exitosamente"},
          "400": {"description": "Datos inválidos o email/username ya existe"}
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Autenticacion"],
        "summary": "Iniciar sesión",
        "description": "Inicia sesión con email y password, retorna token JWT",
        "consumes": ["application/json"],
        "parameters": [{
          "in": "body",
          "name": "body",
          "required": true,
          "schema": {
            "type": "object",
            "required": ["email", "password"],
            "properties": {
              "email": {"type": "string"},
              "password": {"type": "string"}
            }
          }
        }],
        "responses": {
          "200": {"description": "Inicio de sesión exitoso"},
          "401": {"description": "Credenciales inválidas"}
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Autenticacion"],
        "summary": "Cerrar sesión",
        "security": [{"Bearer": []}],
        "responses": {"200": {"description": "Sesión cerrada exitosamente"}}
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Autenticacion"],
        "summary": "Obtener usuario actual",
        "security": [{"Bearer": []}],
        "responses": {"200": {"description": "Datos del usuario actual"}}
      }
    },
...
    "/comments": {
      "get": {
        "tags": ["Comentarios"],
        "summary": "Obtener todos los comentarios",
        "parameters": [
          {"name": "video_id", "in": "query", "type": "integer"},
          {"name": "page", "in": "query", "type": "integer"}
        ],
        "responses": {"200": {"description": "Lista de comentarios"}}
      }
    },
    "/comments/{comment_id}": {
      "get": {
        "tags": ["Comentarios"],
        "summary": "Obtener comentario por ID",
        "parameters": [{"name": "comment_id", "in": "path", "required": true, "type": "integer"}],
        "responses": {"200": {"description": "Comentario encontrado"}}
      },
      "put": {
        "tags": ["Comentarios"],
        "summary": "Actualizar comentario",
        "security": [{"Bearer": []}],
        "parameters": [{"name": "comment_id", "in": "path", "required": true, "type": "integer"}],
        "responses": {"200": {"description": "Comentario actualizado"}, "403": {"description": "No autorizado"}}
      },
      "delete": {
        "tags": ["Comentarios"],
        "summary": "Eliminar comentario",
        "security": [{"Bearer": []}],
        "parameters": [{"name": "comment_id", "in": "path", "required": true, "type": "integer"}],
        "responses": {"200": {"description": "Comentario eliminado"}, "403": {"description": "No autorizado"}}
      }
    },
    "/videos/{video_id}/comments": {
      "post": {
        "tags": ["Comentarios"],
        "summary": "Crear comentario",
        "security": [{"Bearer": []}],
        "parameters": [{"name": "video_id", "in": "path", "required": true, "type": "integer"}],
        "responses": {"201": {"description": "Comentario creado"}, "404": {"description": "Video no encontrado"}}
      }
    },
...
  }
}